zendit/requirements.txt
uvicorn==0.30.1
python-dotenv==1.0.1
fastapi==0.111.0
langchain-community==0.2.11
lancedb==0.11.0